Crate substrate_consensus_slots

Slots functionality for Substrate.

Some consensus algorithms have a concept of slots, which are intervals in time during which certain events can and/or must occur. This crate provides generic functionality for slots.

Structs

SlotDuration
A slot duration. Create with get_or_compute.

SlotInfo
Information about a slot.

Slots
A stream that returns every time there is a new slot.

Enums

CheckedHeader
A header which has been checked

Constants

MAX_SLOT_CAPACITY
We keep at least this number of slots in database.

PRUNING_BOUND
We prune slots when they reach this number.

Traits

SlotCompatible
Slot compatible inherent data.

SlotData
A type from which a slot duration can be obtained.

SlotWorker
A worker that should be invoked at every new slot.

Functions

check_equivocation
Checks if the header is an equivocation and returns the proof in that case.

slot_now
Get the slot for now.

start_slot_worker
Start a new slot worker.